using base64 to convert. showing the image but crushing after click on add.
