+++ DogCare/DogCare.csproj
@@ -46,6 +46,7 @@
     <Compile Include="Pages\WalkDetailPage.xaml.cs">
       <DependentUpon>WalkDetailPage.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Utils\ImageStream.cs" />
     <Compile Include="Utils\ObservableGroupCollection.cs" />
     <Compile Include="Pages\ChangePasswordPage.xaml.cs">
       <DependentUpon>ChangePasswordPage.xaml</DependentUpon>
